feat: insert the selected result into current-post-input

+++ src/app/shared/components/user-list-item/user-list-item.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { UserInterface } from '../../models/user.interface';
 import { AuthService } from '../../../services/auth.service';
 import { ChatService } from '../../../services/chat.service';
@@ -22,11 +22,12 @@ export class UserListItemComponent {
   // Input property that receives a user object from the parent component
   @Input() user!: UserInterface;
   @Input() relatedToSearchResultPost: boolean = false;
-  @Input() inCurrentPostInput = false;
+  @Input() inSearchResultsCurrentPostInput: boolean = false;
   @Input() doNothing: boolean = false;
   @Input() showProfile: boolean = false;
   @Input() inHeaderChat: boolean = false;
 
+  @Output() userSelected = new EventEmitter<UserInterface>();
   // Stores the ID of the currently logged-in user
   currentUserId: string | null = null;
 
@@ -52,14 +53,17 @@ export class UserListItemComponent {
   }
 
   removeFocusAndHandleClick() {
-    // Fokus sofort entfernen, bevor der Klick verarbeitet wird
-    this.searchService.removeFocus();
-    this.choiceBetweenNavigateAndProfile();
+    if (!this.inSearchResultsCurrentPostInput) {
+      // Fokus sofort entfernen, bevor der Klick verarbeitet wird
+      this.searchService.removeFocus();
+      this.choiceBetweenNavigateAndProfile();
+    }
   }
   async choiceBetweenNavigateAndProfile() {
-    if (!this.currentUserId) return; // Stop if user is not logged in
-    if (this.doNothing) return;
-    if (this.showProfile || this.inHeaderChat) {
+    if (!this.currentUserId || this.doNothing) return;
+    if (this.inSearchResultsCurrentPostInput) {
+      this.userSelected.emit(this.user);
+    } else if (this.showProfile || this.inHeaderChat) {
       this.openProfileOverlay();
     } else {
       this.pickOutAndNavigateToChat();
